src/actions/charity_donation/connector/giving_block.py
import logging
import os
from typing import Dict, List, Optional
import requests
from dotenv import load_dotenv
from web3 import Web3
from actions.base import ActionConfig, ActionConnector
from actions.charity_donation.interface import CharityDonationInput

class GivingBlockConnector(ActionConnector[CharityDonationInput]):
    """Connector for charity donations with REAL crypto payments and verification."""

    def __init__(self, config: ActionConfig):
        super().__init__(config)
        load_dotenv()

        rpc_url = os.getenv("ETH_RPC_URL", "https://eth.llamarpc.com")
        self.w3 = Web3(Web3.HTTPProvider(rpc_url))
        
        self.charity_wallet = os.getenv("CHARITY_WALLET_ADDRESS", "0xAb166A6d162Cb2D240bFba55fE9E864FEeC46935")
        
        # Updated charity catalog - Credit Gifting focused
        self.charities = [
            {
                "id": "credit_gift",
                "name": "OpenMind API Credit Gift Program",
                "description": "Gift API credits to developers who have exhausted their free tier",
                "category": "Technology & Education",
                "impact": "$10 = 10,000 API credits for 1 developer",
                "wallet_eth": self.charity_wallet,
                "wallet_usdc": self.charity_wallet,
                "website": "https://github.com/OpenMind/OM1"
            },
            {
                "id": "student_credits",
                "name": "Student Developer Credit Fund",
                "description": "Provide free API access to students learning AI and robotics",
                "category": "Education",
                "impact": "$25 = 1 student gets 6 months of free API access",
                "wallet_eth": self.charity_wallet,
                "wallet_usdc": self.charity_wallet,
                "website": "https://github.com/OpenMind/OM1"
            }
        ]

        self.donations = []
        self.verified_txs = set()

        logging.info("✅ Charity Donation Connector initialized")
        logging.info(f"🌐 Connected to Ethereum: {self.w3.is_connected()}")
        logging.info(f"💳 Charity Wallet: {self.charity_wallet}")
        logging.info(f"❤️  {len(self.charities)} credit gift programs available")

    async def connect(self, output_interface: CharityDonationInput) -> None:
        """Execute charity donation action."""
        try:
            action = output_interface.action.lower()

            if action == "browse_charities":
                await self._browse_charities(output_interface)
            elif action == "donate":
                await self._make_donation(output_interface)
            elif action == "verify_transaction":
                await self._verify_transaction(output_interface)
            elif action == "check_donation":
                await self._check_donation(output_interface)
            else:
                logging.warning(f"Unknown charity action: {action}")

        except Exception as e:
            logging.error(f"Failed to execute charity action: {str(e)}")
            raise

    async def _browse_charities(self, interface: CharityDonationInput) -> None:
        """Browse available charities."""
        try:
            logging.info("\n❤️  API CREDIT GIFT PROGRAMS - REAL CRYPTO DONATIONS\n")
            logging.info("All donations go to REAL wallet addresses on Ethereum blockchain\n")
            
            for charity in self.charities:
                logging.info(f"💝 {charity['name']}")
                logging.info(f"   Category: {charity['category']}")
                logging.info(f"   {charity['description']}")
                logging.info(f"   Impact: {charity['impact']}")
                logging.info(f"   Wallet: {charity['wallet_eth']}")
                logging.info(f"   Verify: https://etherscan.io/address/{charity['wallet_eth']}")
                logging.info("")

        except Exception as e:
            logging.error(f"Failed to browse charities: {str(e)}")
            raise

    async def _make_donation(self, interface: CharityDonationInput) -> None:
        """Make a donation to charity with REAL transaction details."""
        try:
            charity = None
            if interface.charity_id:
                charity = next((c for c in self.charities if c['id'] == interface.charity_id), None)
            elif interface.charity_name:
                search_lower = interface.charity_name.lower()
                charity = next((c for c in self.charities 
                               if search_lower in c['name'].lower() or 
                                  search_lower in c['category'].lower() or
                                  'credit' in search_lower), None)

            if not charity:
                charity = self.charities[0]  # Default to first

            amount = interface.amount_usd or 10.0
            donor_email = interface.donor_email or "anonymous@donor.org"
            eth_price = 2500
            eth_amount = amount / eth_price

            logging.info(f"\n💝 REAL DONATION - Blockchain Transaction")
            logging.info(f"━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
            logging.info(f"Program: {charity['name']}")
            logging.info(f"Amount: ${amount} USD")
            logging.info(f"Your Impact: {charity['impact']}")
            logging.info("")

            import uuid
            donation_id = f"DON-{uuid.uuid4().hex[:8].upper()}"

            logging.info(f"🪙  ETHEREUM PAYMENT (Mainnet or Sepolia)")
            logging.info(f"━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
            logging.info(f"")
            logging.info(f"📍 Recipient Address:")
            logging.info(f"   {charity['wallet_eth']}")
            logging.info(f"")
            logging.info(f"💰 Amount to Send:")
            logging.info(f"   {eth_amount:.6f} ETH (at ~${eth_price}/ETH)")
            logging.info(f"   OR")
            logging.info(f"   {amount:.2f} USDC")
            logging.info(f"")
            logging.info(f"🔗 Verify Address:")
            logging.info(f"   https://etherscan.io/address/{charity['wallet_eth']}")
            logging.info(f"")
            logging.info(f"━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
            logging.info(f"")
            logging.info(f"💡 COMPLETE VIA WEB UI")
            logging.info(f"")
            logging.info(f"📋 Donation ID: {donation_id}")
            logging.info(f"📧 Receipt: {donor_email}")
            logging.info(f"")
            logging.info(f"✨ This is a REAL blockchain transaction!")
            logging.info(f"🙏 Thank you for supporting developers!")

            donation_record = {
                'id': donation_id,
                'charity': charity,
                'amount_usd': amount,
                'amount_eth': eth_amount,
                'donor_email': donor_email,
                'wallet_address': charity['wallet_eth'],
                'status': 'awaiting_payment',
                'tx_hash': None
            }
            self.donations.append(donation_record)

        except Exception as e:
            logging.error(f"Failed to make donation: {str(e)}")
            raise

    async def _verify_transaction(self, interface: CharityDonationInput) -> None:
        """Verify a real blockchain transaction."""
        try:
            tx_hash = interface.donation_id
            
            if not tx_hash or not tx_hash.startswith('0x'):
                logging.error("❌ Invalid transaction hash")
                return

            logging.info(f"\n🔍 VERIFYING BLOCKCHAIN TRANSACTION")
            logging.info(f"━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
            logging.info(f"Transaction Hash: {tx_hash}")

            try:
                tx = self.w3.eth.get_transaction(tx_hash)
                receipt = self.w3.eth.get_transaction_receipt(tx_hash)

                to_address = tx['to']
                value_wei = tx['value']
                value_eth = self.w3.from_wei(value_wei, 'ether')
                
                if to_address and to_address.lower() == self.charity_wallet.lower():
                    status = "✅ SUCCESS" if receipt['status'] == 1 else "❌ FAILED"
                    
                    logging.info(f"Status: {status}")
                    logging.info(f"")
                    logging.info(f"📊 Transaction Details:")
                    logging.info(f"   From: {tx['from']}")
                    logging.info(f"   To: {to_address}")
                    logging.info(f"   Amount: {value_eth} ETH")
                    logging.info(f"   Block: {receipt['blockNumber']}")
                    logging.info(f"")
                    
                    if receipt['status'] == 1:
                        logging.info(f"✅ CREDIT GIFT VERIFIED!")
                        logging.info(f"   Developer credits will be allocated")
                        logging.info(f"   Thank you for supporting the community! ❤️")
                        self.verified_txs.add(tx_hash)

            except Exception:
                logging.error(f"❌ Transaction not found or not yet mined")

        except Exception as e:
            logging.error(f"Failed to verify transaction: {str(e)}")

    async def _check_donation(self, interface: CharityDonationInput) -> None:
        """Check donation status."""
        try:
            donation_id = interface.donation_id
            
            if not donation_id:
                logging.error("Donation ID required")
                return

            donation = next((d for d in self.donations if d['id'] == donation_id), None)

            if not donation:
                logging.error(f"Donation not found: {donation_id}")
                return

            logging.info(f"\n📊 Donation Status")
            logging.info(f"━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━")
            logging.info(f"🆔 Donation ID: {donation['id']}")
            logging.info(f"💝 Program: {donation['charity']['name']}")
            logging.info(f"💰 Amount: ${donation['amount_usd']} USD ({donation['amount_eth']:.6f} ETH)")
            logging.info(f"📧 Email: {donation['donor_email']}")
            logging.info(f"📍 Wallet: {donation['wallet_address']}")
            logging.info(f"✅ Status: {donation['status'].upper()}")

        except Exception as e:
            logging.error(f"Failed to check donation: {str(e)}")
